change iaf_wang_2002 to iaf_bw_2001
