+++ testsuite/regressiontests/ticket-618.sli
@@ -46,7 +46,7 @@ Author: Hans Ekkehard Plesser, 2012-12-11
 
 M_ERROR setverbosity
 
-/excluded_models [ /eprop_iaf_bsshslm_2020 /eprop_iaf_adapt_bsshslm_2020 /eprop_readout_bsshslm_2020 /iaf_wang_2002 ] def
+/excluded_models [ /eprop_iaf_bsshslm_2020 /eprop_iaf_adapt_bsshslm_2020 /eprop_readout_bsshslm_2020 /iaf_bw_2001 ] def
 
 {
   GetKernelStatus /node_models get
